+++ NEWS.md
@@ -3,6 +3,8 @@
 - Adding new `addDataQuality` function for updating data quality matrices in
 soundscape NetCDF files
 
+- Adding more fail catches for `matchGFS`
+
 # PAMscapse 0.16.0
 
 - Adding options `linetype`, `lwd`, `borderLwd`, `showNObs`, `alpha` to `plotPSD`